Stability improvements in extensions and theming Make parent themes and extensions less entangled with process.cwd
