+++ build-packages/scandipwa-dev-utils/package.json
@@ -26,6 +26,7 @@
         "cross-spawn": "^7.0.3",
         "mem-fs": "^1.2.0",
         "mem-fs-editor": "^7.0.1",
+        "memoizee": "^0.4.15",
         "react-dev-utils": "^10.2.1",
         "semver": "^7.3.2",
         "validate-npm-package-name": "^3.0.0"
